+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Added:
-- 
+- Fluid behavior api
 ### Fixed:
 - 
 ### Changed: